# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChainLance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: role
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/suspend:
    patch:
      operationId: suspendUser
      tags: [users]
      summary: Suspend/unsuspend user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # FREELANCERS
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services/gigs
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: freelancerId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: freelancerId
          in: query
          schema: { type: integer }
        - name: serviceId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # NOTIFICATIONS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string }
        avatarUrl: { type: ["string", "null"] }
        walletAddress: { type: ["string", "null"] }
        isSuspended: { type: boolean }
        createdAt: { type: string }
        chainCoinBalance: { type: number }
        level: { type: ["integer", "null"] }
        trustScore: { type: ["number", "null"] }
    UserUpdate:
      type: object
      properties:
        username: { type: string }
        avatarUrl: { type: string }
        walletAddress: { type: string }
    SuspendInput:
      type: object
      required: [suspended]
      properties:
        suspended: { type: boolean }
        reason: { type: string }
    UserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # FREELANCER
    Freelancer:
      type: object
      required: [id, userId, bio, skills, category, hourlyRate, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        user:
          $ref: "#/components/schemas/User"
        bio: { type: string }
        skills:
          type: array
          items: { type: string }
        category: { type: string }
        hourlyRate: { type: number }
        status: { type: string }
        portfolioItems:
          type: array
          items: { type: string }
        completedOrders: { type: integer }
        avgRating: { type: ["number", "null"] }
        totalEarnings: { type: number }
        level: { type: ["integer", "null"] }
        createdAt: { type: string }
...
    Service:
      type: object
      required: [id, freelancerId, title, description, category, price, deliveryDays, status, createdAt]
      properties:
        id: { type: integer }
        freelancerId: { type: integer }
        freelancer:
          $ref: "#/components/schemas/Freelancer"
        title: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        deliveryDays: { type: integer }
        status: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        previewImages:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        totalOrders: { type: integer }
        avgRating: { type: ["number", "null"] }
        isFeatured: { type: boolean }
        createdAt: { type: string }
    ServiceInput:
      type: object
      required: [title, description, category, price, deliveryDays]
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        deliveryDays: { type: integer }
        thumbnailUrl: { type: string }
        previewImages:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
    ServiceUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        deliveryDays: { type: integer }
        thumbnailUrl: { type: string }
        previewImages:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string }
    ServiceList:
      type: object
      required: [services, total, page, limit]
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ORDER
...
    Review:
      type: object
      required: [id, orderId, reviewerId, freelancerId, rating, comment, createdAt]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        reviewerId: { type: integer }
        reviewer:
          $ref: "#/components/schemas/User"
        freelancerId: { type: integer }
        serviceId: { type: ["integer", "null"] }
        rating: { type: integer }
        comment: { type: string }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [orderId, rating, comment]
      properties:
        orderId: { type: integer }
        rating: { type: integer }
        comment: { type: string }

    # NOTIFICATIONS
...